backend/requirements.txt
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-dotenv==1.2.1
google-generativeai==0.8.6
requests==2.32.5
reportlab==3.6.8
pypdf==6.0.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
